+++ src/QuantumClifford.jl
@@ -83,9 +83,7 @@ export
     # makie plotting -- defined only when extension is loaded
     stabilizerplot, stabilizerplot_axis,
     # sum types
-    compactify_circuit,
-    # qoptics
-    stab_to_ket
+    compactify_circuit
 
 
 const BIG_INT_MINUS_ONE = Ref{BigInt}()
@@ -1168,6 +1166,5 @@ include("precompiles.jl")
 include("ecc/ECC.jl")
 include("nonclifford.jl")
 include("plotting_extensions.jl")
-include("qoptics_extensions.jl")
 
 end #module